espresso

+++ admin/founditem_d.php
@@ -3,15 +3,19 @@
 require '../database/connect_db.php';
 
 $ItemID = $_POST['ItemID'];
-if(isset($ItemID)){
-$erase = mysqli_query($con, "DELETE FROM foundItems_t WHERE ItemID = $ItemID");
+$check = mysqli_query($con, "SELECT $ItemID FROM foundItems_t WHERE ItemID = $ItemID");
+
+if(!empty($check)){
+  $erase = mysqli_query($con, "DELETE FROM foundItems_t WHERE ItemID = $ItemID");
 }
+
 if($erase){
     Header('Refresh: 3; admin.php');
     echo "Found item entry was deleted successfully.";
-
 }
+
 else{
     echo "Unable to execute $erase. ". mysqli_error($con);
   }
+
  ?>
